requirements.txt
discord.py
python-decouple
pytz
python-dateutil
Unidecode